Correct the rates should be expressed in MultiArrhenius form within FFCM mechanism
There are rates of total 7 reactions in FFCM should be shown in FFCM form, but the previous version has only one with that form.
The rest 6 only has single Arrhenius form corresponding to the first Arrhenius in FFCM.
Now they are corrected and should be able to capture both high and low temperature.